# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Carve API
  description: |
    The Carve language-learning API. This document is the source of truth for
    every `/v1` endpoint; the TypeScript client in `apps/web/src/lib/api.ts`
    is regenerated from it and schemathesis runs property-based HTTP fuzz
    against it as part of the L4 testing layer.
  version: 0.3.0
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.carve.app
    description: Production
security: []  # default: no auth; protected operations explicitly set `security: [{ bearer: [] }]`
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LanguageCode:
      type: string
      enum: [ja, zh-cn, zh-tw, zh, ko, en]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    ImportResultResponse:
      type: object
      properties:
        imported:    { type: integer, minimum: 0 }
        skipped:     { type: integer, minimum: 0 }
        deck_id:     { type: string, nullable: true }
        deck_name:   { type: string, nullable: true }
        scheduling_preserved: { type: boolean }
        warnings:
          type: array
          items: { type: string }

    # ── Output ────────────────────────────────────────────────────────────
...
paths:
...
  /v1/import/anki:
    post:
      tags: [import]
      operationId: importAnki
      summary: Import an Anki .apkg archive (preserves SM-2 → FSRS scheduling)
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:     { type: string, format: binary }
                deck_id:  { type: string }
                language: { $ref: '#/components/schemas/LanguageCode' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ImportResultResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/import/migaku-csv:
    post:
      tags: [import]
      operationId: importMigakuCSV
      summary: Import a Migaku CSV export
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:     { type: string, format: binary }
                language: { $ref: '#/components/schemas/LanguageCode' }
                deck_id:  { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ImportResultResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/import/yomitan:
    post:
      tags: [import]
      operationId: importYomitan
      summary: Import a Yomitan vocabulary export
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:     { type: string, format: binary }
                language: { $ref: '#/components/schemas/LanguageCode' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ImportResultResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /v1/import/jpdb-csv:
    post:
      tags: [import]
      operationId: importJPDBCSV
      summary: Import a jpdb.io vocabulary CSV
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ImportResultResponse' } } } }
        '400': { description: Bad request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  # ── Output ────────────────────────────────────────────────────────────
...